requirements.txt
asgiref==3.5.0
crispy-bootstrap5==0.6
Django==3.2
django-ajax-selects==2.1.0
django-crispy-forms==1.14.0
Pillow==9.0.1
pytz==2021.3
sqlparse==0.4.2
